Modularization. Modern string formatting. Use argparse for cleaner CLI arguments parsing.
