+++ fahrplan/meta.py
@@ -4,4 +4,6 @@ version = '0.2.3'
 description = 'A SBB/CFF/FFS commandline based timetable client.'
 author = 'Danilo Bargen'
 author_email = 'gezuru@gmail.com'
+url = 'https://github.com/dbrgn/fahrplan'
+keywords = 'fahrplan timetable sbb cff ffs public transport'
 license = 'GPLv3'
