+++ setup.py
@@ -14,12 +14,12 @@ setup(name='fahrplan',
       description=meta.description,
       author=meta.author,
       author_email=meta.author_email,
-      url='https://github.com/dbrgn/fahrplan',
+      url=meta.url,
       packages=find_packages(),
       zip_safe=False,
       include_package_data=True,
       license=meta.license,
-      keywords='fahrplan timetable sbb cff ffs public transport',
+      keywords=meta.keywords,
       long_description=readme,
       install_requires=requirements,
       entry_points={
